backend/pdf-to-graphical-abstract/graphical-abstract-generator/src/visualization/visualizer.py
class Visualizer:
    def create_charts(self, data):
        import matplotlib.pyplot as plt
        import seaborn as sns
        
        sns.set_theme(style="whitegrid")
        plt.figure(figsize=(10, 6))
        sns.barplot(x='category', y='value', data=data)
        plt.title('Key Findings Visualization')
        plt.xlabel('Categories')
        plt.ylabel('Values')
        plt.show()

    def overlay_text(self, image_path, text, position=(10, 10)):
        from PIL import Image, ImageDraw, ImageFont
        import cv2
        
        image = Image.open(image_path)
        draw = ImageDraw.Draw(image)
        font = ImageFont.load_default()
        draw.text(position, text, fill="black", font=font)
        image.save('output_image.png')

    def generate_illustration(self, prompt):
        import requests
        
        # Example for DALL·E API call
        api_url = "https://api.openai.com/v1/images/generations"
        headers = {
            "Authorization": "Bearer YOUR_API_KEY",
            "Content-Type": "application/json"
        }
        data = {
            "prompt": prompt,
            "n": 1,
            "size": "1024x1024"
        }
        response = requests.post(api_url, headers=headers, json=data)
        return response.json()