backend/pdf-to-graphical-abstract/graphical-abstract-generator/requirements.txt
matplotlib==3.5.1
HuggingFace==4.21.1
seaborn==0.11.2
Pillow==9.0.1
opencv-python==4.5.3
requests==2.26.0
# Add any necessary API libraries for DALL·E or Stable Diffusion here